+++ site-cookbooks/my-nginx/attributes/default.rb
@@ -1,4 +1,0 @@
-default['my-nginx']['servername']       = node['hostname']
-default['my-nginx']['worker_processes'] = node[:cpu][:total]
-default['my-nginx']['default_root']     = '/var/www/'
-default['my-nginx']['app']              = 'app'