+++ site-cookbooks/my_nginx/attributes/default.rb
@@ -1,0 +1,4 @@
+default['my_nginx']['servername']       = node['hostname']
+default['my_nginx']['worker_processes'] = node[:cpu][:total]
+default['my_nginx']['default_root']     = '/var/www/'
+default['my_nginx']['app']              = 'app'