+++ site-cookbooks/my_mysql/CHANGELOG.md
@@ -1,11 +1,11 @@
-my-mysql CHANGELOG
+my_mysql CHANGELOG
 ==================
 
-This file is used to list changes made in each version of the my-mysql cookbook.
+This file is used to list changes made in each version of the my_mysql cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of my-mysql
+- [your_name] - Initial release of my_mysql
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.