+++ site-cookbooks/my_nginx/CHANGELOG.md
@@ -1,11 +1,11 @@
-my-nginx CHANGELOG
+my_nginx CHANGELOG
 ==================
 
-This file is used to list changes made in each version of the my-nginx cookbook.
+This file is used to list changes made in each version of the my_nginx cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of my-nginx
+- [your_name] - Initial release of my_nginx
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.